+++ application/src/ext/java/org/opentripplanner/ext/vehiclerentalservicedirectory/VehicleRentalServiceDirectoryFetcher.java
@@ -19,6 +19,8 @@ import org.opentripplanner.ext.vehiclerentalservicedirectory.api.VehicleRentalSe
 import org.opentripplanner.framework.io.HttpHeaders;
 import org.opentripplanner.framework.io.OtpHttpClientException;
 import org.opentripplanner.framework.io.OtpHttpClientFactory;
+import org.opentripplanner.gbfs.network.GbfsNetworkOverrides;
+import org.opentripplanner.gbfs.network.GeofencingZonePhase;
 import org.opentripplanner.service.vehiclerental.VehicleRentalRepository;
 import org.opentripplanner.street.linking.VertexLinker;
 import org.opentripplanner.updater.spi.GraphUpdater;
@@ -60,6 +62,7 @@ public class VehicleRentalServiceDirectoryFetcher {
 
   public static List<GraphUpdater<?>> createUpdatersFromEndpoint(
     VehicleRentalServiceDirectoryFetcherParameters parameters,
+    GbfsNetworkOverrides overrides,
     VertexLinker vertexLinker,
     VehicleRentalRepository repository
   ) {
@@ -82,20 +85,22 @@ public class VehicleRentalServiceDirectoryFetcher {
       repository,
       otpHttpClientFactory
     );
-    return serviceDirectory.createUpdatersFromManifest(parameters, manifest);
+    return serviceDirectory.createUpdatersFromManifest(parameters, overrides, manifest);
   }
 
   public List<GraphUpdater<?>> createUpdatersFromManifest(
     VehicleRentalServiceDirectoryFetcherParameters parameters,
+    GbfsNetworkOverrides overrides,
     GBFSManifest manifest
   ) {
     return fetchUpdaterInfoFromDirectoryAndCreateUpdaters(
-      buildListOfNetworksFromManifest(parameters, manifest)
+      buildListOfNetworksFromManifest(parameters, overrides, manifest)
     );
   }
 
   private static List<GbfsVehicleRentalDataSourceParameters> buildListOfNetworksFromManifest(
     VehicleRentalServiceDirectoryFetcherParameters parameters,
+    GbfsNetworkOverrides overrides,
     GBFSManifest manifest
   ) {
     List<GbfsVehicleRentalDataSourceParameters> dataSources = new ArrayList<>();
@@ -109,29 +114,29 @@ public class VehicleRentalServiceDirectoryFetcher {
         continue;
       }
 
-      var config = parameters.networkParameters(networkName);
+      var config = overrides.forNetwork(networkName);
 
-      if (config.isPresent()) {
-        var networkParams = config.get();
-        dataSources.add(
-          new GbfsVehicleRentalDataSourceParameters(
-            gbfsUrl.get(),
-            parameters.getLanguage(),
-            networkParams.allowKeepingAtDestination(),
-            parameters.getHeaders(),
-            networkName,
-            networkParams.geofencingZones(),
-            // geofencingBusinessAreaBorders - default to true
-            true,
-            // overloadingAllowed - not part of GBFS, not supported here
-            false,
-            // rentalPickupType not supported
-            RentalPickupType.ALL
-          )
-        );
-      } else {
+      if (config.isEmpty()) {
         LOG.warn("Network not configured in OTP: {}", networkName);
+        continue;
       }
+
+      var networkParams = config.get();
+      dataSources.add(
+        new GbfsVehicleRentalDataSourceParameters(
+          gbfsUrl.get(),
+          parameters.getLanguage(),
+          networkParams.allowKeepingVehicleAtDestination(),
+          parameters.getHeaders(),
+          networkName,
+          networkParams.geofencingZones() == GeofencingZonePhase.REALTIME,
+          networkParams.requireDropOffInsideBusinessArea(),
+          // overloadingAllowed - not part of GBFS, not supported here
+          false,
+          // rentalPickupType not supported
+          RentalPickupType.ALL
+        )
+      );
     }
     return dataSources;
   }