+++ doc/user/Changelog.md
@@ -5,6 +5,8 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 
 ## 2.10.0-SNAPSHOT (in progress)
 
+- Move per-network GBFS configuration into a shared `gbfs` section of `otp-config.json`
+
 - Add select/not filters to estimatedCalls in Transmodel [#7305](https://github.com/opentripplanner/OpenTripPlanner/pull/7305)
 - Cleanup regular transfer build parameters [#7425](https://github.com/opentripplanner/OpenTripPlanner/pull/7425)
 - Add CostlyAssertions feature that should be turned off in production and improve performance of via routing with coordinates [#7301](https://github.com/opentripplanner/OpenTripPlanner/pull/7301)